fixes, deleted testing route
