--- webapp/src/routes/Board [testing]/BoardContext.jsx
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const BoardContext = createContext();
-
-export default BoardContext;
