--- webapp/src/routes/Board [testing]/ChessContext.jsx
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const ChessContext = createContext();
-
-export default ChessContext;
