--- webapp/src/routes/Board [testing]/style.css
@@ -1,130 +1,0 @@
-/* Placeholder */
-.wrapper {
-	width: 1200px;
-	display: flex;
-	flex-direction: row;
-	justify-content: space-between;
-}
-/* Placeholder */
-
-/* Placeholder */
-.board-container {
-	width: 360px;
-	height: 360px;
-	margin: 20px;
-}
-/* Placeholder */
-
-.board {
-	width: 100%;
-	height: 100%;
-	display: grid;
-	grid-template: repeat(8, 1fr) / repeat(8, 1fr);
-	background-color: #ebecd0;
-}
-
-.box {
-	min-width: 100%;
-	max-width: 100%;
-	min-height: 100%;
-	max-height: 100%;
-}
-
-.box:nth-child(-2n + 8),
-.box:nth-child(8) ~ div:nth-child(-2n + 15),
-.box:nth-child(16) ~ div:nth-child(-2n + 24),
-.box:nth-child(24) ~ div:nth-child(-2n + 31),
-.box:nth-child(32) ~ div:nth-child(-2n + 40),
-.box:nth-child(40) ~ div:nth-child(-2n + 47),
-.box:nth-child(48) ~ div:nth-child(-2n + 56),
-.box:nth-child(56) ~ div:nth-child(-2n + 63) {
-	background-color: #779556;
-}
-
-.selected-white {
-	background-color: #f6f48b;
-}
-
-.selected-black {
-	background-color: #bbca47;
-}
-
-.box img {
-	position: relative;
-	top: -200%;
-	z-index: 50;
-	width: 90%;
-}
-
-.box .overlay {
-	position: relative;
-	z-index: 100;
-	top: -100%;
-	width: 100%;
-	height: 100%;
-}
-
-.box .underlay {
-	position: relative;
-	z-index: 1;
-	top: 0%;
-	width: 100%;
-	height: 100%;
-}
-
-.box .idle {
-	width: 60%;
-	z-index: 2;
-	position: relative;
-	top: 29%;
-	left: 9%;
-}
-
-.box .targeted {
-	width: 139%;
-	z-index: 2;
-	position: relative;
-	top: 0%;
-	left: 0%;
-}
-
-/* default skins */
-
-.box .default-bishop {
-	margin-left: 20%;
-	margin-top: 10%;
-}
-
-.box .default-pawn {
-	margin-left: 25%;
-	margin-top: 20%;
-}
-
-.box .default-rook {
-	margin-left: 22%;
-	margin-top: 19%;
-}
-
-.box .default-knight {
-	margin-left: 15%;
-	margin-top: 12%;
-}
-
-.box .default-queen {
-	margin-left: 10%;
-	margin-top: 12%;
-}
-
-.box .default-king {
-	margin-left: 11.5%;
-	margin-top: 10%;
-}
-
-.box .wikipedia-bishop, .wikipedia-king, .wikipedia-knight, .wikipedia-pawn, .wikipedia-queen, .wikipedia-rook {
-	width: 100%;
-	margin-top: 2%;
-}
-
-.box .wikipedia-bishop {
-	margin-top: -3%;
-}
